- fix: specfile queries with BuildArch: (#27589).
CVS patchset: 4746
CVS date: 2001/05/06 19:17:14

+++ lib/fsm.c
@@ -1126,7 +1126,7 @@ int fsmStage(FSM_t fsm, fileStage stage)
     struct stat * ost = &fsm->osb;
     int saveerrno = errno;
     int rc = fsm->rc;
-    int left;
+    size_t left;
     int i;
 
 #define	_fafilter(_a)	\
